main page in another folder
I change it and put the main page in another folder that it has the name of Main page
